requirements.txt
asgiref==3.8.1
aspose-words==25.3.0
attrs==25.3.0
dj-rest-auth==7.0.1
Django==5.2
django-bootstrap5==25.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
Markdown==3.8
numpy==2.2.4
pandas==2.2.3
pillow==11.2.1
Pygments==2.19.1
sqlparse==0.5.3
typing_extensions==4.13.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
pillow==11.2.1
drf-yasg==1.26.0
PyJWT==2.9.0
django-rest-swagger==2.2.0
django-cors-headers==4.7.0